+++ src/cases_generation/generate_one.py
@@ -5,10 +5,10 @@ import numpy as np
 from pathlib import Path
 from typing import Tuple
 import warnings
-from utils import torch_layer
+from src.utils.utils import torch_layer
 from functools import partial
 
-from utils import get_layer_func
+from src.utils.utils import get_layer_func
 
 warnings.filterwarnings("ignore")
