+++ src/cases_generation/tool.py
@@ -1,6 +1,7 @@
 import numpy as np
 import onnxruntime
 from onnx import numpy_helper
+import torch
 
 
 # some helper functions to check weight
@@ -8,6 +9,7 @@ from onnx import numpy_helper
 # convert tensor to numpy
 def to_numpy(tensor: torch.Tensor):
     return tensor.detach().cpu().numpy() if tensor.requires_grad else tensor.cpu().numpy()
+
 
 # differential testing
 # at this stage assume tensor1 and tensor2 are both numpy array
@@ -19,12 +21,15 @@ def diff_test(tensor1,tensor2,thres = 10e-1,norm = "L2"):
     if norm == "Linf":
         return np.max(np.abs(tensor1-tensor2)) <= thres
 
+
 def extract_pytorch_weight(model):
     return [to_numpy(para) for para in model.parameters()]
+
 
 def extract_ort_weight(model):
     weights = model.graph.initializer
     return [numpy_helper.to_array(weights[i]) for i in range(len(weights))]
+
 
 def ort_inference(onnx_path, input):
     ort_session = onnxruntime.InferenceSession(onnx_path)
